feat(home): 合并横屏与竖屏主页组件
